[hotfix] give rule classes more descriptive names and refactor to separate files
