+++ .gitignore
@@ -17,3 +17,4 @@
 ################################################################################
 
 venv
+__pycache__
